cleanup p2
