+++ src/classes/combiner/__init__.py
@@ -1,0 +1,5 @@
+from .combiner import Combiner
+
+__all__ = [
+    "Combiner",
+]
