+++ src/classes/pack/__init__.py
@@ -1,0 +1,5 @@
+from .pack import Pack
+
+__all__ = [
+    "Pack",
+]
